WebGL 3D ! Coord3D from int to uint8_t Lights header more offsets
UI
==
- Monitor: change to new header structure. WebGL: add gl-matrix

Server
=====
- Package.json: add gl-matrix (for 3D webgl in monitor) ˜100KB in WWWData.h
- Utilities: Coord3D from int to uint8_t!!! (12 bytes to 3 bytes), add intToCoord3D (WIP)
- Effects and Layouts and Modifiers: into to uint8_t
- Nodes: addLight: uint16_t to uint8_t
- Physical layer: refactor LightsHeader: add more offSets
- Virtual layer: drawLine from int to uint8_t

+++ src/MoonLight/PhysicalLayer.cpp
@@ -71,7 +71,7 @@ PhysicalLayer::PhysicalLayer() {
                 //send the positions to the UI _socket_emit
                 //reset the index and continue...
             } else
-                lights.positions[lights.header.nrOfLights] = {(uint16_t)position.x, (uint16_t)position.y, (uint16_t)position.z};
+                lights.positions[lights.header.nrOfLights] = {position.x, position.y, position.z};
             
             lights.header.size = lights.header.size.maximum(position);
         } else {
